Move duplicate code into parent class
